refactor: split test.cc into per-subsystem test files
Split the monolithic 1538-line test.cc into 6 focused files:
- test_containers.cc: FixedArray, DynArray, Dictionary, CircularBuffer,
  SegmentedList, InlinedArray, BlockAllocator, FreeList
- test_strings.cc: StringBuffer, NullTerminated, StringTable
- test_math.cc: Vectors, Bits, Easing
- test_error.cc: Error, ErrorOr, TRY, MUST
- test_collision.cc: narrow-phase Collision, CollisionWorld
- test_executor.cc: InlineExecutor, ThreadPoolExecutor

Add test_fixture.h with AllocTest base class that provides a system
allocator, reducing boilerplate across tests that need one.
